Display data from Job and Company on the Homepage, edit the route for the Dashboard Component, use Routes to display the content of each tab in the Sidebar based on the endpoint URL.
