+++ src/apis/job.api.ts
@@ -1,10 +1,9 @@
 import http from '../utils/http.ts';
 import { AxiosRequestConfig } from 'axios';
-import { JobList, JobListConfig, JobResponse } from '../types/jobResponse.type.ts';
+import { JobList, JobListConfig, JobResponse } from '../types/job.type.ts';
 
 const jobApi = {
   getAllJobs: (params: JobListConfig & AxiosRequestConfig) => {
-    console.log('hehe: ', params);
     return http.get<JobList>('/jobs', { params });
   },
   getJobDetail: (id: string) => http.get<JobResponse>(`/jobs/${id}`)
