+++ src/components/FeatureJob/FeatureJob.tsx
@@ -2,36 +2,22 @@ import { Link } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
 import useQueryParams from '../../hooks/useQueryPrams.tsx';
 import jobApi from '../../apis/job.api.ts';
+import { calcDayRemaining, formatSalary } from '../../utils/utils.ts';
 // import { getAccessTokenFromLocalStorage } from '../../utils/auth.ts';
 // import axios from 'axios';
 
 export default function FeatureJob() {
+  const { VITE_API_BASE_URL } = import.meta.env;
+
   const queryParams = useQueryParams();
 
-  const { data } = useQuery({
+  const { data: jobData } = useQuery({
     queryKey: ['JobList', queryParams],
     queryFn: () => {
-      console.log('params: ', queryParams);
       return jobApi.getAllJobs(queryParams);
     }
   });
-  console.log('check list job: ', data);
 
-  // const fetchData = async () => {
-  //   try {
-  //     const response = await axios.get('http://localhost:8080/api/v1/jobs', {
-  //       params: queryParams,
-  //       headers: {
-  //         Authorization: `Bearer ${getAccessTokenFromLocalStorage()}`
-  //       }
-  //     });
-  //     console.log(response.data);
-  //   } catch (error) {
-  //     console.error(error);
-  //   }
-  // };
-
-  // fetchData();
   return (
     <div className='bg-white my-[100px]'>
       <div className='container'>
@@ -70,988 +56,178 @@ export default function FeatureJob() {
           </Link>
         </div>
         <div className='mt-[50px]'>
-          <div
-            className='flex justify-between items-center p-[32px] border border-[#EDEFF5] cursor-pointer
+          {jobData &&
+            jobData.data.data.data.slice(0, 6).map((job) => (
+              <div
+                key={job.id}
+                className='flex justify-between items-center p-[32px] border border-[#EDEFF5] cursor-pointer
              rounded-[12px] hover:shadow-2xl hover:transition hover:ease-in-out hover:duration-[0.25s]'
-          >
-            <div className='flex items-center gap-[20px]'>
-              <img
-                src='src/assets/images/tiktok.png'
-                alt=''
-                className='w-[68px] h-[68px] object-cover rounded-[6px]'
-              />
+              >
+                <div className='flex items-center gap-[20px]'>
+                  <img
+                    src={`${VITE_API_BASE_URL}/files/${job.company.logo?.fileName}`}
+                    alt=''
+                    className='w-[68px] h-[68px] object-cover rounded-[6px]'
+                  />
 
-              <div>
-                <Link to='/'>
-                  <h3 className='text-[20px] leading-8 font-medium text-[#191F33]'>
-                    Senior UX Design
-                  </h3>
-                </Link>
+                  <div>
+                    <Link to='/'>
+                      <h3 className='text-[20px] leading-8 font-medium text-[#191F33]'>
+                        {job.name}
+                      </h3>
+                    </Link>
 
-                <div className='flex items-center gap-[16px] mt-[14px]'>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <path
-                        d='M19.25 9.16675C19.25 15.5834 11 21.0834 11 21.0834C11 21.0834 2.75 15.5834 2.75 9.16675C2.75 6.97871 3.61919 4.88029 5.16637 3.33312C6.71354 1.78594 8.81196 0.916748 11 0.916748C13.188 0.916748 15.2865 1.78594 16.8336 3.33312C18.3808 4.88029 19.25 6.97871 19.25 9.16675Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                      <path
-                        d='M11 11.9167C12.5188 11.9167 13.75 10.6855 13.75 9.16675C13.75 7.64797 12.5188 6.41675 11 6.41675C9.48122 6.41675 8.25 7.64797 8.25 9.16675C8.25 10.6855 9.48122 11.9167 11 11.9167Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                    </svg>
+                    <div className='flex items-center gap-[16px] mt-[14px]'>
+                      <div className='flex items-center gap-[6px]'>
+                        <svg
+                          width='22'
+                          height='22'
+                          viewBox='0 0 22 22'
+                          fill='none'
+                          xmlns='http://www.w3.org/2000/svg'
+                        >
+                          <path
+                            d='M19.25 9.16675C19.25 15.5834 11 21.0834 11 21.0834C11 21.0834 2.75 15.5834 2.75 9.16675C2.75 6.97871 3.61919 4.88029 5.16637 3.33312C6.71354 1.78594 8.81196 0.916748 11 0.916748C13.188 0.916748 15.2865 1.78594 16.8336 3.33312C18.3808 4.88029 19.25 6.97871 19.25 9.16675Z'
+                            stroke='#C5C9D6'
+                            strokeWidth='1.5'
+                            strokeLinecap='round'
+                            strokeLinejoin='round'
+                          />
+                          <path
+                            d='M11 11.9167C12.5188 11.9167 13.75 10.6855 13.75 9.16675C13.75 7.64797 12.5188 6.41675 11 6.41675C9.48122 6.41675 8.25 7.64797 8.25 9.16675C8.25 10.6855 9.48122 11.9167 11 11.9167Z'
+                            stroke='#C5C9D6'
+                            strokeWidth='1.5'
+                            strokeLinecap='round'
+                            strokeLinejoin='round'
+                          />
+                        </svg>
 
-                    <div className='text-[14px] leading-5 text-[#636A80]'>Australia</div>
+                        <div className='text-[14px] leading-5 text-[#636A80]'>{job.location}</div>
+                      </div>
+                      <div className='flex items-center gap-[4px]'>
+                        <svg
+                          width='22'
+                          height='22'
+                          viewBox='0 0 22 22'
+                          fill='none'
+                          xmlns='http://www.w3.org/2000/svg'
+                        >
+                          <g clipPath='url(#clip0_1647_32162)'>
+                            <path
+                              d='M11 2.0625V19.9375'
+                              stroke='#C5C9D6'
+                              strokeWidth='1.5'
+                              strokeLinecap='round'
+                              strokeLinejoin='round'
+                            />
+                            <path
+                              d='M15.8125 7.5625C15.8125 7.11108 15.7236 6.66408 15.5508 6.24703C15.3781 5.82997 15.1249 5.45102 14.8057 5.13182C14.4865 4.81262 14.1075 4.55941 13.6905 4.38666C13.2734 4.21391 12.8264 4.125 12.375 4.125H9.28125C8.36957 4.125 7.49523 4.48716 6.85057 5.13182C6.20591 5.77648 5.84375 6.65082 5.84375 7.5625C5.84375 8.47418 6.20591 9.34852 6.85057 9.99318C7.49523 10.6378 8.36957 11 9.28125 11H13.0625C13.9742 11 14.8485 11.3622 15.4932 12.0068C16.1378 12.6515 16.5 13.5258 16.5 14.4375C16.5 15.3492 16.1378 16.2235 15.4932 16.8682C14.8485 17.5128 13.9742 17.875 13.0625 17.875H8.9375C8.02582 17.875 7.15148 17.5128 6.50682 16.8682C5.86216 16.2235 5.5 15.3492 5.5 14.4375'
+                              stroke='#C5C9D6'
+                              strokeWidth='1.5'
+                              strokeLinecap='round'
+                              strokeLinejoin='round'
+                            />
+                          </g>
+                          <defs>
+                            <clipPath id='clip0_1647_32162'>
+                              <rect width='22' height='22' fill='white' />
+                            </clipPath>
+                          </defs>
+                        </svg>
+
+                        <div className='text-[14px] leading-5 text-[#636A80]'>
+                          {formatSalary(job.salary)}
+                        </div>
+                      </div>
+                      <div className='flex items-center gap-[6px]'>
+                        <svg
+                          width='22'
+                          height='22'
+                          viewBox='0 0 22 22'
+                          fill='none'
+                          xmlns='http://www.w3.org/2000/svg'
+                        >
+                          <g clipPath='url(#clip0_1647_32168)'>
+                            <path
+                              d='M17.875 3.4375H4.125C3.7453 3.4375 3.4375 3.7453 3.4375 4.125V17.875C3.4375 18.2547 3.7453 18.5625 4.125 18.5625H17.875C18.2547 18.5625 18.5625 18.2547 18.5625 17.875V4.125C18.5625 3.7453 18.2547 3.4375 17.875 3.4375Z'
+                              stroke='#C5C9D6'
+                              strokeWidth='1.5'
+                              strokeLinecap='round'
+                              strokeLinejoin='round'
+                            />
+                            <path
+                              d='M15.125 2.0625V4.8125'
+                              stroke='#C5C9D6'
+                              strokeWidth='1.5'
+                              strokeLinecap='round'
+                              strokeLinejoin='round'
+                            />
+                            <path
+                              d='M6.875 2.0625V4.8125'
+                              stroke='#C5C9D6'
+                              strokeWidth='1.5'
+                              strokeLinecap='round'
+                              strokeLinejoin='round'
+                            />
+                            <path
+                              d='M3.4375 7.5625H18.5625'
+                              stroke='#C5C9D6'
+                              strokeWidth='1.5'
+                              strokeLinecap='round'
+                              strokeLinejoin='round'
+                            />
+                          </g>
+                          <defs>
+                            <clipPath id='clip0_1647_32168'>
+                              <rect width='22' height='22' fill='white' />
+                            </clipPath>
+                          </defs>
+                        </svg>
+
+                        <div className='text-[14px] leading-5 text-[#636A80]'>
+                          {calcDayRemaining(job.endDate) > 1
+                            ? calcDayRemaining(job.endDate) + ' Days Remaining'
+                            : calcDayRemaining(job.endDate) + ' Day Remaining'}
+                        </div>
+                      </div>
+                    </div>
                   </div>
-                  <div className='flex items-center gap-[4px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32162)'>
+                </div>
+                <div>
+                  <Link
+                    className='flex items-center gap-[12px] bg-[#E7F0FA] px-6 py-3 rounded-[12px]
+              text-[16px] font-semibold text-[#0A65CC]'
+                    to='/'
+                  >
+                    Apply Now{' '}
+                    <div>
+                      <svg
+                        width='24'
+                        height='24'
+                        viewBox='0 0 24 24'
+                        fill='none'
+                        xmlns='http://www.w3.org/2000/svg'
+                      >
                         <path
-                          d='M11 2.0625V19.9375'
-                          stroke='#C5C9D6'
+                          d='M5 12H19'
+                          stroke='#0A65CC'
                           strokeWidth='1.5'
                           strokeLinecap='round'
                           strokeLinejoin='round'
                         />
                         <path
-                          d='M15.8125 7.5625C15.8125 7.11108 15.7236 6.66408 15.5508 6.24703C15.3781 5.82997 15.1249 5.45102 14.8057 5.13182C14.4865 4.81262 14.1075 4.55941 13.6905 4.38666C13.2734 4.21391 12.8264 4.125 12.375 4.125H9.28125C8.36957 4.125 7.49523 4.48716 6.85057 5.13182C6.20591 5.77648 5.84375 6.65082 5.84375 7.5625C5.84375 8.47418 6.20591 9.34852 6.85057 9.99318C7.49523 10.6378 8.36957 11 9.28125 11H13.0625C13.9742 11 14.8485 11.3622 15.4932 12.0068C16.1378 12.6515 16.5 13.5258 16.5 14.4375C16.5 15.3492 16.1378 16.2235 15.4932 16.8682C14.8485 17.5128 13.9742 17.875 13.0625 17.875H8.9375C8.02582 17.875 7.15148 17.5128 6.50682 16.8682C5.86216 16.2235 5.5 15.3492 5.5 14.4375'
-                          stroke='#C5C9D6'
+                          d='M12 5L19 12L12 19'
+                          stroke='#0A65CC'
                           strokeWidth='1.5'
                           strokeLinecap='round'
                           strokeLinejoin='round'
                         />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32162'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'> $120,000 - $140,000</div>
-                  </div>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32168)'>
-                        <path
-                          d='M17.875 3.4375H4.125C3.7453 3.4375 3.4375 3.7453 3.4375 4.125V17.875C3.4375 18.2547 3.7453 18.5625 4.125 18.5625H17.875C18.2547 18.5625 18.5625 18.2547 18.5625 17.875V4.125C18.5625 3.7453 18.2547 3.4375 17.875 3.4375Z'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.125 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M6.875 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M3.4375 7.5625H18.5625'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32168'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>4 Days Remaining</div>
-                  </div>
+                      </svg>
+                    </div>
+                  </Link>
                 </div>
               </div>
-            </div>
-            <div>
-              <Link
-                className='flex items-center gap-[12px] bg-[#E7F0FA] px-6 py-3 rounded-[12px]
-              text-[16px] font-semibold text-[#0A65CC]'
-                to='/'
-              >
-                Apply Now{' '}
-                <div>
-                  <svg
-                    width='24'
-                    height='24'
-                    viewBox='0 0 24 24'
-                    fill='none'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path
-                      d='M5 12H19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                    <path
-                      d='M12 5L19 12L12 19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                  </svg>
-                </div>
-              </Link>
-            </div>
-          </div>
-
-          <div
-            className='flex justify-between items-center p-[32px] border border-[#EDEFF5]
-            cursor-pointer rounded-[12px] hover:shadow-2xl hover:transition hover:ease-in-out
-            hover:duration-[0.25s] mt-[24px]'
-          >
-            <div className='flex items-center gap-[20px]'>
-              <img
-                src='src/assets/images/tiktok.png'
-                alt=''
-                className='w-[68px] h-[68px] object-cover rounded-[6px]'
-              />
-
-              <div>
-                <Link to='/'>
-                  <h3 className='text-[20px] leading-8 font-medium text-[#191F33]'>
-                    Senior UX Design
-                  </h3>
-                </Link>
-
-                <div className='flex items-center gap-[16px] mt-[14px]'>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <path
-                        d='M19.25 9.16675C19.25 15.5834 11 21.0834 11 21.0834C11 21.0834 2.75 15.5834 2.75 9.16675C2.75 6.97871 3.61919 4.88029 5.16637 3.33312C6.71354 1.78594 8.81196 0.916748 11 0.916748C13.188 0.916748 15.2865 1.78594 16.8336 3.33312C18.3808 4.88029 19.25 6.97871 19.25 9.16675Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                      <path
-                        d='M11 11.9167C12.5188 11.9167 13.75 10.6855 13.75 9.16675C13.75 7.64797 12.5188 6.41675 11 6.41675C9.48122 6.41675 8.25 7.64797 8.25 9.16675C8.25 10.6855 9.48122 11.9167 11 11.9167Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>Australia</div>
-                  </div>
-                  <div className='flex items-center gap-[4px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32162)'>
-                        <path
-                          d='M11 2.0625V19.9375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.8125 7.5625C15.8125 7.11108 15.7236 6.66408 15.5508 6.24703C15.3781 5.82997 15.1249 5.45102 14.8057 5.13182C14.4865 4.81262 14.1075 4.55941 13.6905 4.38666C13.2734 4.21391 12.8264 4.125 12.375 4.125H9.28125C8.36957 4.125 7.49523 4.48716 6.85057 5.13182C6.20591 5.77648 5.84375 6.65082 5.84375 7.5625C5.84375 8.47418 6.20591 9.34852 6.85057 9.99318C7.49523 10.6378 8.36957 11 9.28125 11H13.0625C13.9742 11 14.8485 11.3622 15.4932 12.0068C16.1378 12.6515 16.5 13.5258 16.5 14.4375C16.5 15.3492 16.1378 16.2235 15.4932 16.8682C14.8485 17.5128 13.9742 17.875 13.0625 17.875H8.9375C8.02582 17.875 7.15148 17.5128 6.50682 16.8682C5.86216 16.2235 5.5 15.3492 5.5 14.4375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32162'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'> $120,000 - $140,000</div>
-                  </div>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32168)'>
-                        <path
-                          d='M17.875 3.4375H4.125C3.7453 3.4375 3.4375 3.7453 3.4375 4.125V17.875C3.4375 18.2547 3.7453 18.5625 4.125 18.5625H17.875C18.2547 18.5625 18.5625 18.2547 18.5625 17.875V4.125C18.5625 3.7453 18.2547 3.4375 17.875 3.4375Z'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.125 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M6.875 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M3.4375 7.5625H18.5625'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32168'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>4 Days Remaining</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div>
-              <Link
-                className='flex items-center gap-[12px] bg-[#E7F0FA] px-6 py-3 rounded-[12px]
-              text-[16px] font-semibold text-[#0A65CC]'
-                to='/'
-              >
-                Apply Now{' '}
-                <div>
-                  <svg
-                    width='24'
-                    height='24'
-                    viewBox='0 0 24 24'
-                    fill='none'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path
-                      d='M5 12H19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                    <path
-                      d='M12 5L19 12L12 19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                  </svg>
-                </div>
-              </Link>
-            </div>
-          </div>
-
-          <div
-            className='flex justify-between items-center p-[32px] border border-[#EDEFF5]
-            cursor-pointer rounded-[12px] hover:shadow-2xl hover:transition hover:ease-in-out
-            hover:duration-[0.25s] mt-[24px]'
-          >
-            <div className='flex items-center gap-[20px]'>
-              <img
-                src='src/assets/images/tiktok.png'
-                alt=''
-                className='w-[68px] h-[68px] object-cover rounded-[6px]'
-              />
-
-              <div>
-                <Link to='/'>
-                  <h3 className='text-[20px] leading-8 font-medium text-[#191F33]'>
-                    Senior UX Design
-                  </h3>
-                </Link>
-
-                <div className='flex items-center gap-[16px] mt-[14px]'>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <path
-                        d='M19.25 9.16675C19.25 15.5834 11 21.0834 11 21.0834C11 21.0834 2.75 15.5834 2.75 9.16675C2.75 6.97871 3.61919 4.88029 5.16637 3.33312C6.71354 1.78594 8.81196 0.916748 11 0.916748C13.188 0.916748 15.2865 1.78594 16.8336 3.33312C18.3808 4.88029 19.25 6.97871 19.25 9.16675Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                      <path
-                        d='M11 11.9167C12.5188 11.9167 13.75 10.6855 13.75 9.16675C13.75 7.64797 12.5188 6.41675 11 6.41675C9.48122 6.41675 8.25 7.64797 8.25 9.16675C8.25 10.6855 9.48122 11.9167 11 11.9167Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>Australia</div>
-                  </div>
-                  <div className='flex items-center gap-[4px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32162)'>
-                        <path
-                          d='M11 2.0625V19.9375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.8125 7.5625C15.8125 7.11108 15.7236 6.66408 15.5508 6.24703C15.3781 5.82997 15.1249 5.45102 14.8057 5.13182C14.4865 4.81262 14.1075 4.55941 13.6905 4.38666C13.2734 4.21391 12.8264 4.125 12.375 4.125H9.28125C8.36957 4.125 7.49523 4.48716 6.85057 5.13182C6.20591 5.77648 5.84375 6.65082 5.84375 7.5625C5.84375 8.47418 6.20591 9.34852 6.85057 9.99318C7.49523 10.6378 8.36957 11 9.28125 11H13.0625C13.9742 11 14.8485 11.3622 15.4932 12.0068C16.1378 12.6515 16.5 13.5258 16.5 14.4375C16.5 15.3492 16.1378 16.2235 15.4932 16.8682C14.8485 17.5128 13.9742 17.875 13.0625 17.875H8.9375C8.02582 17.875 7.15148 17.5128 6.50682 16.8682C5.86216 16.2235 5.5 15.3492 5.5 14.4375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32162'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'> $120,000 - $140,000</div>
-                  </div>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32168)'>
-                        <path
-                          d='M17.875 3.4375H4.125C3.7453 3.4375 3.4375 3.7453 3.4375 4.125V17.875C3.4375 18.2547 3.7453 18.5625 4.125 18.5625H17.875C18.2547 18.5625 18.5625 18.2547 18.5625 17.875V4.125C18.5625 3.7453 18.2547 3.4375 17.875 3.4375Z'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.125 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M6.875 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M3.4375 7.5625H18.5625'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32168'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>4 Days Remaining</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div>
-              <Link
-                className='flex items-center gap-[12px] bg-[#E7F0FA] px-6 py-3 rounded-[12px]
-              text-[16px] font-semibold text-[#0A65CC]'
-                to='/'
-              >
-                Apply Now{' '}
-                <div>
-                  <svg
-                    width='24'
-                    height='24'
-                    viewBox='0 0 24 24'
-                    fill='none'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path
-                      d='M5 12H19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                    <path
-                      d='M12 5L19 12L12 19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                  </svg>
-                </div>
-              </Link>
-            </div>
-          </div>
-
-          <div
-            className='flex justify-between items-center p-[32px] border border-[#EDEFF5]
-            cursor-pointer rounded-[12px] hover:shadow-2xl hover:transition hover:ease-in-out
-            hover:duration-[0.25s] mt-[24px]'
-          >
-            <div className='flex items-center gap-[20px]'>
-              <img
-                src='src/assets/images/tiktok.png'
-                alt=''
-                className='w-[68px] h-[68px] object-cover rounded-[6px]'
-              />
-
-              <div>
-                <Link to='/'>
-                  <h3 className='text-[20px] leading-8 font-medium text-[#191F33]'>
-                    Senior UX Design
-                  </h3>
-                </Link>
-
-                <div className='flex items-center gap-[16px] mt-[14px]'>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <path
-                        d='M19.25 9.16675C19.25 15.5834 11 21.0834 11 21.0834C11 21.0834 2.75 15.5834 2.75 9.16675C2.75 6.97871 3.61919 4.88029 5.16637 3.33312C6.71354 1.78594 8.81196 0.916748 11 0.916748C13.188 0.916748 15.2865 1.78594 16.8336 3.33312C18.3808 4.88029 19.25 6.97871 19.25 9.16675Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                      <path
-                        d='M11 11.9167C12.5188 11.9167 13.75 10.6855 13.75 9.16675C13.75 7.64797 12.5188 6.41675 11 6.41675C9.48122 6.41675 8.25 7.64797 8.25 9.16675C8.25 10.6855 9.48122 11.9167 11 11.9167Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>Australia</div>
-                  </div>
-                  <div className='flex items-center gap-[4px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32162)'>
-                        <path
-                          d='M11 2.0625V19.9375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.8125 7.5625C15.8125 7.11108 15.7236 6.66408 15.5508 6.24703C15.3781 5.82997 15.1249 5.45102 14.8057 5.13182C14.4865 4.81262 14.1075 4.55941 13.6905 4.38666C13.2734 4.21391 12.8264 4.125 12.375 4.125H9.28125C8.36957 4.125 7.49523 4.48716 6.85057 5.13182C6.20591 5.77648 5.84375 6.65082 5.84375 7.5625C5.84375 8.47418 6.20591 9.34852 6.85057 9.99318C7.49523 10.6378 8.36957 11 9.28125 11H13.0625C13.9742 11 14.8485 11.3622 15.4932 12.0068C16.1378 12.6515 16.5 13.5258 16.5 14.4375C16.5 15.3492 16.1378 16.2235 15.4932 16.8682C14.8485 17.5128 13.9742 17.875 13.0625 17.875H8.9375C8.02582 17.875 7.15148 17.5128 6.50682 16.8682C5.86216 16.2235 5.5 15.3492 5.5 14.4375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32162'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'> $120,000 - $140,000</div>
-                  </div>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32168)'>
-                        <path
-                          d='M17.875 3.4375H4.125C3.7453 3.4375 3.4375 3.7453 3.4375 4.125V17.875C3.4375 18.2547 3.7453 18.5625 4.125 18.5625H17.875C18.2547 18.5625 18.5625 18.2547 18.5625 17.875V4.125C18.5625 3.7453 18.2547 3.4375 17.875 3.4375Z'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.125 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M6.875 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M3.4375 7.5625H18.5625'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32168'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>4 Days Remaining</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div>
-              <Link
-                className='flex items-center gap-[12px] bg-[#E7F0FA] px-6 py-3 rounded-[12px]
-              text-[16px] font-semibold text-[#0A65CC]'
-                to='/'
-              >
-                Apply Now{' '}
-                <div>
-                  <svg
-                    width='24'
-                    height='24'
-                    viewBox='0 0 24 24'
-                    fill='none'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path
-                      d='M5 12H19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                    <path
-                      d='M12 5L19 12L12 19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                  </svg>
-                </div>
-              </Link>
-            </div>
-          </div>
-
-          <div
-            className='flex justify-between items-center p-[32px] border border-[#EDEFF5]
-            cursor-pointer rounded-[12px] hover:shadow-2xl hover:transition hover:ease-in-out
-            hover:duration-[0.25s] mt-[24px]'
-          >
-            <div className='flex items-center gap-[20px]'>
-              <img
-                src='src/assets/images/tiktok.png'
-                alt=''
-                className='w-[68px] h-[68px] object-cover rounded-[6px]'
-              />
-
-              <div>
-                <Link to='/'>
-                  <h3 className='text-[20px] leading-8 font-medium text-[#191F33]'>
-                    Senior UX Design
-                  </h3>
-                </Link>
-
-                <div className='flex items-center gap-[16px] mt-[14px]'>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <path
-                        d='M19.25 9.16675C19.25 15.5834 11 21.0834 11 21.0834C11 21.0834 2.75 15.5834 2.75 9.16675C2.75 6.97871 3.61919 4.88029 5.16637 3.33312C6.71354 1.78594 8.81196 0.916748 11 0.916748C13.188 0.916748 15.2865 1.78594 16.8336 3.33312C18.3808 4.88029 19.25 6.97871 19.25 9.16675Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                      <path
-                        d='M11 11.9167C12.5188 11.9167 13.75 10.6855 13.75 9.16675C13.75 7.64797 12.5188 6.41675 11 6.41675C9.48122 6.41675 8.25 7.64797 8.25 9.16675C8.25 10.6855 9.48122 11.9167 11 11.9167Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>Australia</div>
-                  </div>
-                  <div className='flex items-center gap-[4px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32162)'>
-                        <path
-                          d='M11 2.0625V19.9375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.8125 7.5625C15.8125 7.11108 15.7236 6.66408 15.5508 6.24703C15.3781 5.82997 15.1249 5.45102 14.8057 5.13182C14.4865 4.81262 14.1075 4.55941 13.6905 4.38666C13.2734 4.21391 12.8264 4.125 12.375 4.125H9.28125C8.36957 4.125 7.49523 4.48716 6.85057 5.13182C6.20591 5.77648 5.84375 6.65082 5.84375 7.5625C5.84375 8.47418 6.20591 9.34852 6.85057 9.99318C7.49523 10.6378 8.36957 11 9.28125 11H13.0625C13.9742 11 14.8485 11.3622 15.4932 12.0068C16.1378 12.6515 16.5 13.5258 16.5 14.4375C16.5 15.3492 16.1378 16.2235 15.4932 16.8682C14.8485 17.5128 13.9742 17.875 13.0625 17.875H8.9375C8.02582 17.875 7.15148 17.5128 6.50682 16.8682C5.86216 16.2235 5.5 15.3492 5.5 14.4375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32162'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'> $120,000 - $140,000</div>
-                  </div>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32168)'>
-                        <path
-                          d='M17.875 3.4375H4.125C3.7453 3.4375 3.4375 3.7453 3.4375 4.125V17.875C3.4375 18.2547 3.7453 18.5625 4.125 18.5625H17.875C18.2547 18.5625 18.5625 18.2547 18.5625 17.875V4.125C18.5625 3.7453 18.2547 3.4375 17.875 3.4375Z'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.125 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M6.875 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M3.4375 7.5625H18.5625'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32168'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>4 Days Remaining</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div>
-              <Link
-                className='flex items-center gap-[12px] bg-[#E7F0FA] px-6 py-3 rounded-[12px]
-              text-[16px] font-semibold text-[#0A65CC]'
-                to='/'
-              >
-                Apply Now{' '}
-                <div>
-                  <svg
-                    width='24'
-                    height='24'
-                    viewBox='0 0 24 24'
-                    fill='none'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path
-                      d='M5 12H19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                    <path
-                      d='M12 5L19 12L12 19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                  </svg>
-                </div>
-              </Link>
-            </div>
-          </div>
-
-          <div
-            className='flex justify-between items-center p-[32px] border border-[#EDEFF5]
-            cursor-pointer rounded-[12px] hover:shadow-2xl hover:transition hover:ease-in-out
-            hover:duration-[0.25s] mt-[24px]'
-          >
-            <div className='flex items-center gap-[20px]'>
-              <img
-                src='src/assets/images/tiktok.png'
-                alt=''
-                className='w-[68px] h-[68px] object-cover rounded-[6px]'
-              />
-
-              <div>
-                <Link to='/'>
-                  <h3 className='text-[20px] leading-8 font-medium text-[#191F33]'>
-                    Senior UX Design
-                  </h3>
-                </Link>
-
-                <div className='flex items-center gap-[16px] mt-[14px]'>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <path
-                        d='M19.25 9.16675C19.25 15.5834 11 21.0834 11 21.0834C11 21.0834 2.75 15.5834 2.75 9.16675C2.75 6.97871 3.61919 4.88029 5.16637 3.33312C6.71354 1.78594 8.81196 0.916748 11 0.916748C13.188 0.916748 15.2865 1.78594 16.8336 3.33312C18.3808 4.88029 19.25 6.97871 19.25 9.16675Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                      <path
-                        d='M11 11.9167C12.5188 11.9167 13.75 10.6855 13.75 9.16675C13.75 7.64797 12.5188 6.41675 11 6.41675C9.48122 6.41675 8.25 7.64797 8.25 9.16675C8.25 10.6855 9.48122 11.9167 11 11.9167Z'
-                        stroke='#C5C9D6'
-                        strokeWidth='1.5'
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                      />
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>Australia</div>
-                  </div>
-                  <div className='flex items-center gap-[4px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32162)'>
-                        <path
-                          d='M11 2.0625V19.9375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.8125 7.5625C15.8125 7.11108 15.7236 6.66408 15.5508 6.24703C15.3781 5.82997 15.1249 5.45102 14.8057 5.13182C14.4865 4.81262 14.1075 4.55941 13.6905 4.38666C13.2734 4.21391 12.8264 4.125 12.375 4.125H9.28125C8.36957 4.125 7.49523 4.48716 6.85057 5.13182C6.20591 5.77648 5.84375 6.65082 5.84375 7.5625C5.84375 8.47418 6.20591 9.34852 6.85057 9.99318C7.49523 10.6378 8.36957 11 9.28125 11H13.0625C13.9742 11 14.8485 11.3622 15.4932 12.0068C16.1378 12.6515 16.5 13.5258 16.5 14.4375C16.5 15.3492 16.1378 16.2235 15.4932 16.8682C14.8485 17.5128 13.9742 17.875 13.0625 17.875H8.9375C8.02582 17.875 7.15148 17.5128 6.50682 16.8682C5.86216 16.2235 5.5 15.3492 5.5 14.4375'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32162'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'> $120,000 - $140,000</div>
-                  </div>
-                  <div className='flex items-center gap-[6px]'>
-                    <svg
-                      width='22'
-                      height='22'
-                      viewBox='0 0 22 22'
-                      fill='none'
-                      xmlns='http://www.w3.org/2000/svg'
-                    >
-                      <g clipPath='url(#clip0_1647_32168)'>
-                        <path
-                          d='M17.875 3.4375H4.125C3.7453 3.4375 3.4375 3.7453 3.4375 4.125V17.875C3.4375 18.2547 3.7453 18.5625 4.125 18.5625H17.875C18.2547 18.5625 18.5625 18.2547 18.5625 17.875V4.125C18.5625 3.7453 18.2547 3.4375 17.875 3.4375Z'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M15.125 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M6.875 2.0625V4.8125'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                        <path
-                          d='M3.4375 7.5625H18.5625'
-                          stroke='#C5C9D6'
-                          strokeWidth='1.5'
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                        />
-                      </g>
-                      <defs>
-                        <clipPath id='clip0_1647_32168'>
-                          <rect width='22' height='22' fill='white' />
-                        </clipPath>
-                      </defs>
-                    </svg>
-
-                    <div className='text-[14px] leading-5 text-[#636A80]'>4 Days Remaining</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div>
-              <Link
-                className='flex items-center gap-[12px] bg-[#E7F0FA] px-6 py-3 rounded-[12px]
-              text-[16px] font-semibold text-[#0A65CC]'
-                to='/'
-              >
-                Apply Now{' '}
-                <div>
-                  <svg
-                    width='24'
-                    height='24'
-                    viewBox='0 0 24 24'
-                    fill='none'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path
-                      d='M5 12H19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                    <path
-                      d='M12 5L19 12L12 19'
-                      stroke='#0A65CC'
-                      strokeWidth='1.5'
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                    />
-                  </svg>
-                </div>
-              </Link>
-            </div>
-          </div>
+            ))}
         </div>
       </div>
     </div>
