--- src/pages/AdminDashboard/AdminDashboard.tsx
@@ -1,3 +1,0 @@
-export default function AdminDashboard() {
-  return <div>Admin Dashboard</div>;
-}
