--- src/pages/AdminDashboard/index.ts
@@ -1,3 +1,0 @@
-import AdminDashboard from './AdminDashboard.tsx';
-
-export default AdminDashboard;
