+++ src/pages/Dashboard/Company/Company.tsx
@@ -1,0 +1,3 @@
+export default function Company() {
+  return <div className='mt-[54px]'>Company</div>;
+}
