+++ src/pages/Dashboard/Company/index.ts
@@ -1,0 +1,3 @@
+import Company from './Company.tsx';
+
+export default Company;
