+++ src/pages/Dashboard/Overview/index.ts
@@ -1,0 +1,3 @@
+import Overview from './Overview.tsx';
+
+export default Overview;
