+++ src/pages/Dashboard/SidebarDashboard/index.ts
@@ -1,0 +1,3 @@
+import SidebarDashboard from './SidebarDashboard.tsx';
+
+export default SidebarDashboard;
