--- src/pages/HrDashboard/HrDashboard.tsx
@@ -1,3 +1,0 @@
-export default function HrDashboard() {
-  return <div>HR Dashboard</div>;
-}
