--- src/pages/HrDashboard/index.ts
@@ -1,3 +1,0 @@
-import HrDashboard from './HrDashboard.tsx';
-
-export default HrDashboard;
