--- src/pages/UserDashboard/index.ts
@@ -1,3 +1,0 @@
-import UserDashboard from './UserDashboard.tsx';
-
-export default UserDashboard;
