+++ src/types/company.type.ts
@@ -1,11 +1,18 @@
 import { ByUser } from './user.type.ts';
+import { MetaData } from './meta.type.ts';
 
 export interface Company {
   id: string;
   name: string;
   description: string;
   address: string;
-  logo: string | null;
+  logo: {
+    id: string;
+    fileName: string;
+    fileType: string;
+    size: number;
+    uploadTime: string;
+  } | null;
   user: ByUser[] | [];
   createdBy: ByUser | null;
   createdAt: string | null;
@@ -14,4 +21,26 @@ export interface Company {
   deletedBy: ByUser | null;
   deletedAt: string | null;
   deleted: boolean;
+}
+
+export interface CompanyResponse {
+  message: string;
+  statusCode: number;
+  data: Company;
+}
+
+export interface CompanyList {
+  message: string;
+  statusCode: number;
+  data: {
+    meta: MetaData;
+    data: Company[] | [];
+  };
+}
+
+export interface CompanyConfig {
+  pageNo?: number | string;
+  pageSize?: number | string;
+  sortBy?: number | string;
+  sortDir?: number | string;
 }
