+++ src/types/job.type.ts
@@ -1,6 +1,7 @@
 import { ByUser } from './user.type.ts';
 import { Company } from './company.type.ts';
 import { Skill } from './skill.type.ts';
+import { MetaData } from './meta.type.ts';
 
 export interface Job {
   id: string;
@@ -17,9 +18,31 @@ export interface Job {
   deletedBy: ByUser | null;
   deletedAt: string | null;
   company: Company;
-  startDate: string | null;
-  endDate: string | null;
+  startDate: string;
+  endDate: string;
   skills: Skill[] | [];
   active: boolean;
   deleted: boolean;
+}
+
+export interface JobResponse {
+  message: string;
+  statusCode: number;
+  data: Job;
+}
+
+export interface JobList {
+  message: string;
+  statusCode: number;
+  data: {
+    meta: MetaData;
+    data: Job[] | [];
+  };
+}
+
+export interface JobListConfig {
+  pageNo?: number | string;
+  pageSize?: number | string;
+  sortBy?: number | string;
+  sortDir?: number | string;
 }
